+++ app/lib/shared/versions.dart
@@ -22,7 +22,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.08.16',
+  '2023.08.17',
   // Fallback runtime versions.
   '2023.08.08',
   '2023.08.04',
@@ -54,10 +54,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.0.0';
-final String toolStableDartSdkVersion = '3.0.7';
-final String toolStableFlutterSdkVersion = '3.10.6';
-final String toolPreviewDartSdkVersion = '3.1.0-262.3.beta';
-final String toolPreviewFlutterSdkVersion = '3.13.0-0.3.pre';
+final String toolStableDartSdkVersion = '3.1.0';
+final String toolStableFlutterSdkVersion = '3.13.0';
+final String toolPreviewDartSdkVersion = '3.2.0-42.1.beta';
+final String toolPreviewFlutterSdkVersion = '3.13.0';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);