+++ CHANGELOG.md
@@ -2,7 +2,11 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2023.08.16`.
+ * Bumped runtimeVersion to `2023.08.17`.
+ * Upgraded stable Dart analysis SDK to `3.1.0`.
+ * Upgraded stable Flutter analysis SDK to `3.13.0`.
+ * Upgraded preview Dart analysis SDK to `3.2.0-42.1.beta`.
+ * Upgraded preview Flutter analysis SDK to `3.13.0`.
  * Note: Upgraded `package:gcloud`.
  * Note: `analyzer` and `dartdoc` processing is turned off.
  * Note: `search` has now a separate index in an isolate that is renewed every 15 minutes.